+++ Bosses/Lych/MiniLychConfig.cs
@@ -35,8 +35,8 @@ public static class MiniLychConfig {
     public static void ApplyMiniLychSettings(BloonModel bloonModel, string difficulty, int level) {
         float multiplier = difficultyMultipliers[difficulty];
 
-        bloonModel.maxHealth = baseMaxHealth * (levelMaxHealthMultiplier * level) * multiplier;
-        if (level == 0) { bloonModel.maxHealth = baseMaxHealth * multiplier; }
+        bloonModel.maxHealth = (int)(baseMaxHealth * (levelMaxHealthMultiplier * level) * multiplier);
+        if (level == 0) { bloonModel.maxHealth = (int)(baseMaxHealth * multiplier); }
 
         bloonModel.leakDamage = 99999f;
         bloonModel.speed = (baseSpeed + levelSpeedAddition * level) * multiplier;
